refactor(useactionslogic): moved logic to separate hooks based on pr feedback
