+++ apps/backoffice-v2/src/pages/Entity/components/Case/hooks/useActions/interfaces.ts
@@ -1,4 +1,4 @@
-export interface IUseActions {
+export interface IUseActionsLogic {
   workflowId: string;
   fullName: string;
 }
